ml-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
prophet>=1.1.5
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.6.0
joblib>=1.3.2
python-multipart>=0.0.6
sqlalchemy>=2.0.25
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0 